blog_app/serializers.py
from rest_framework import serializers
from blog_app.models import BlogPost,Feedback

class BlogPostSerializer(serializers.ModelSerializer):
    class Meta:
        model = BlogPost
        fields = '__all__'
        lookup_field = 'slug'
        
class FeedbackSerializer(serializers.ModelSerializer):
    class Meta:
        model = Feedback
        exclude = ['date_received']
        