requirements.txt
asgiref==3.4.1
bleach==4.1.0
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
django-summernote==0.8.20.0
djangorestframework==3.12.4
packaging==21.0
Pillow==8.4.0
psycopg2==2.9.2
pyparsing==3.0.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
whitenoise==5.3.0
